Add files via upload
Collaborate with the risk-management department to adjust risk measurement metrics and ensure prompt notification to merchants on various scenarios.
This is one of those scenarios: to notify a bunch of merchants about problems related to their URLs. 
Before that, remember to sort out the right merchants to the right account managers.
Good coding tips:
Writing steps before start coding aways.
Test methods with a small amount of data.
When I started to write the same codes/logics over and over again, it is time to write a general method.